+++ app/preload/index.js
@@ -24,16 +24,11 @@ require("./follow")
 
 const {pathToSpecialPageName} = require("../util")
 const specialPage = pathToSpecialPageName(window.location.href)
-const viewerProtocols = ["sourceviewer:", "readerview:", "markdownviewer:"]
+const skipProtocols = ["sourceviewer:", "readerview:", "markdownviewer:"]
 if (specialPage?.name) {
     // Load the special page specific JavaScript
     require(`./${specialPage.name}`)
-} else if (viewerProtocols.some(p => window.location.href.startsWith(p))) {
-    // Load the viewer specific functions
-    if (window.location.href.startsWith("markdownviewer:")) {
-        require("./markdownviewer")
-    }
-} else {
+} else if (!skipProtocols.some(p => window.location.href.startsWith(p))) {
     // Load the failed page information handler for nonspecial pages
     require("./failedload")
     // Load the local directory browser for nonspecial pages