+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 - Failedload and filebrowser pages are now styled with themes like other special pages
 - Undo/redo navbar logic to prevent duplicate entries in the edit history
+- Commands to list the name and arguments of failing commands more often
 
 ### Fixed